deb/ch4/ep2/requirements.txt
pandas==1.3.5
jupyterlab==3.4.0
PyYAML==6.0
requests_oauthlib==1.3.1
Flask==2.1.2
SQLAlchemy==1.4.36
PyMySQL==1.0.2